+++ lib/tabbed_view.dart
@@ -15,6 +15,7 @@ export 'src/tabbed_view_icons.dart';
 export 'src/tabbed_view_menu_builder.dart';
 export 'src/tabbed_view_menu_item.dart';
 export 'src/theme/content_area_theme_data.dart';
+export 'src/theme/last_visible_tab_behavior.dart';
 export 'src/theme/side_tabs_layout.dart';
 export 'src/theme/tab_decoration_builder.dart';
 export 'src/theme/tab_header_extent_behavior.dart';